# event_go/internal/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Soulmarker Event API",
    "version": "1.0.0",
    "description": "Versioned API contract for authentication, organization authorization, organizers, events, registrations, admissions, checkins, discussions and tickets. Legacy /api routes remain available for one compatibility cycle."
  },
  "servers": [{ "url": "/api/v1" }],
...
  "paths": {
...
    "/organizations": {
      "post": {
        "operationId": "createOrganization",
        "tags": ["Organizations"],
        "summary": "Create an organization workspace owned by the current user",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/CreateOrganizationRequest" },
        "responses": {
          "201": { "$ref": "#/components/responses/Created" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/organization-invitations/accept": {
      "post": {
        "operationId": "acceptOrganizationInvitation",
        "tags": ["Organizations"],
        "summary": "Accept an organization invitation as the current user",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/AcceptOrganizationInvitationRequest" },
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
    "/organizations/{organizationId}": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "get": {
        "operationId": "getOrganizationWorkspace",
        "tags": ["Organizations"],
        "summary": "Get the current tenant workspace and organizer profile",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/organizations/{organizationId}/members": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "get": {
        "operationId": "listOrganizationMembers",
        "tags": ["Organizations"],
        "summary": "List members visible to the current tenant principal",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/organizations/{organizationId}/members/{memberId}": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }, { "$ref": "#/components/parameters/OrganizationMemberID" }],
      "put": {
        "operationId": "updateOrganizationMember",
        "tags": ["Organizations"],
        "summary": "Change an organization member role within tenant policy",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/UpdateOrganizationMemberRequest" },
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "operationId": "revokeOrganizationMember",
        "tags": ["Organizations"],
        "summary": "Revoke an organization member within tenant policy",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/organizations/{organizationId}/owner-transfer": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "post": {
        "operationId": "transferOrganizationOwnership",
        "tags": ["Organizations"],
        "summary": "Transfer organization ownership to another active member",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/TransferOrganizationOwnerRequest" },
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/organizations/{organizationId}/invitations": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "get": {
        "operationId": "listOrganizationInvitations",
        "tags": ["Organizations"],
        "summary": "List organization invitations without secret token material",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "operationId": "createOrganizationInvitation",
        "tags": ["Organizations"],
        "summary": "Invite a user to the organization and deliver a one-time link",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/CreateOrganizationInvitationRequest" },
        "responses": {
          "201": { "$ref": "#/components/responses/Created" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/organizations/{organizationId}/invitations/{invitationId}": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }, { "$ref": "#/components/parameters/OrganizationInvitationID" }],
      "delete": {
        "operationId": "revokeOrganizationInvitation",
        "tags": ["Organizations"],
        "summary": "Revoke a pending organization invitation",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/organizers": {
      "post": {
        "operationId": "createOrganizer",
        "tags": ["Organizers"],
        "summary": "Create an organizer",
        "security": [{ "AdminToken": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/CreateOrganizerRequest" },
        "responses": {
          "201": { "$ref": "#/components/responses/Created" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "get": {
        "operationId": "listOrganizers",
        "tags": ["Organizers"],
        "summary": "List organizers",
        "parameters": [{ "$ref": "#/components/parameters/Page" }, { "$ref": "#/components/parameters/PageSize" }],
        "responses": { "200": { "$ref": "#/components/responses/OK" } }
      }
    },
    "/organizers/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizerID" }],
      "get": {
        "operationId": "getOrganizer",
        "tags": ["Organizers"],
        "summary": "Get an organizer",
        "responses": { "200": { "$ref": "#/components/responses/OK" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "put": {
        "operationId": "updateOrganizer",
        "tags": ["Organizers"],
        "summary": "Update an organizer",
        "security": [{ "AdminToken": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/UpdateOrganizerRequest" },
        "responses": { "200": { "$ref": "#/components/responses/OK" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "delete": {
        "operationId": "deleteOrganizer",
        "tags": ["Organizers"],
        "summary": "Delete an organizer while preserving historical events",
        "security": [{ "AdminToken": [] }],
        "responses": { "200": { "$ref": "#/components/responses/OK" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "AdminToken": { "type": "apiKey", "in": "header", "name": "X-Admin-Token", "description": "Platform administrator credential; it does not grant an organization membership or tenant capability." }
    },
    "parameters": {
      "Page": { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
      "PageSize": { "name": "page_size", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } },
      "OrganizationID": { "name": "organizationId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
      "OrganizationMemberID": { "name": "memberId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
      "OrganizationInvitationID": { "name": "invitationId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
      "OrganizerID": { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
...
    },
    "requestBodies": {
...
      "CreateOrganizationRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateOrganizationRequest" } } } },
      "CreateOrganizationInvitationRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateOrganizationInvitationRequest" } } } },
      "AcceptOrganizationInvitationRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AcceptOrganizationInvitationRequest" } } } },
      "UpdateOrganizationMemberRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateOrganizationMemberRequest" } } } },
      "TransferOrganizationOwnerRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransferOrganizationOwnerRequest" } } } },
      "CreateOrganizerRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateOrganizerRequest" } } } },
      "UpdateOrganizerRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateOrganizerRequest" } } } },
...
    },
    "responses": {
      "OK": { "description": "Successful response", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/APIResponse" } } } },
      "Created": { "description": "Resource created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/APIResponse" } } } },
      "BadRequest": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Unauthorized": { "description": "Authentication failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Forbidden": { "description": "Authenticated but not allowed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "NotFound": { "description": "Resource not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Conflict": { "description": "Business state conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
...
    },
    "schemas": {
      "APIResponse": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": { "type": "integer" },
          "error_code": { "type": "string", "enum": ["ADMIN_AUTH_INVALID", "ADMISSION_ALREADY_CHECKED_IN", "ADMISSION_NOT_FOUND", "ADMISSION_REVOKED", "API_ROUTE_NOT_FOUND", "CANCELLATION_DEADLINE_EXCEEDED", "CONTENT_ALREADY_REMOVED", "CONTENT_ALREADY_VISIBLE", "CONTENT_REPORT_ALREADY_RESOLVED", "CONTENT_REPORT_NOT_ALLOWED", "CONTENT_REPORT_NOT_FOUND", "EVENT_CAPACITY_FULL", "EVENT_HAS_ADMISSIONS", "EVENT_NOT_FOUND", "EVENT_NOT_PUBLISHED", "INTERNAL_ERROR", "INVALID_CREDENTIALS", "INVALID_JSON", "METHOD_NOT_ALLOWED", "NOTIFICATION_NOT_FOUND", "ORGANIZATION_ACCESS_DENIED", "ORGANIZATION_INVITATION_EXISTS", "ORGANIZATION_INVITATION_INVALID", "ORGANIZATION_MEMBER_CHANGE_DENIED", "ORGANIZATION_MEMBER_EXISTS", "ORGANIZATION_MEMBER_NOT_FOUND", "ORGANIZATION_OWNER_TRANSFER_DENIED", "ORGANIZATION_SLUG_IN_USE", "ORGANIZER_NOT_FOUND", "PARTICIPATION_REQUIRED", "PASSWORD_RESET_TOKEN_INVALID", "POST_NOT_FOUND", "RECOVERY_EMAIL_ALREADY_BOUND", "RECOVERY_EMAIL_IN_USE", "RECOVERY_EMAIL_RATE_LIMITED", "RECOVERY_EMAIL_TOKEN_INVALID", "REGISTRATION_DUPLICATE", "REGISTRATION_NOT_FOUND", "REPLY_NOT_FOUND", "REQUEST_TOO_LARGE", "TICKET_NOT_FOUND", "TICKET_SOLD_OUT", "USER_ALREADY_EXISTS", "USER_AUTH_REQUIRED", "USER_TOKEN_INVALID", "VALIDATION_ERROR"] },
          "message": { "type": "string" },
          "data": {},
          "total": { "type": "integer", "minimum": 0 },
          "page": { "type": "integer", "minimum": 1 },
          "page_size": { "type": "integer", "minimum": 1, "maximum": 100 }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "error_code", "message"],
        "properties": {
          "code": { "type": "integer", "minimum": 400, "maximum": 599 },
          "error_code": { "type": "string", "enum": ["ADMIN_AUTH_INVALID", "ADMISSION_ALREADY_CHECKED_IN", "ADMISSION_NOT_FOUND", "ADMISSION_REVOKED", "API_ROUTE_NOT_FOUND", "CANCELLATION_DEADLINE_EXCEEDED", "CONTENT_ALREADY_REMOVED", "CONTENT_ALREADY_VISIBLE", "CONTENT_REPORT_ALREADY_RESOLVED", "CONTENT_REPORT_NOT_ALLOWED", "CONTENT_REPORT_NOT_FOUND", "EVENT_CAPACITY_FULL", "EVENT_HAS_ADMISSIONS", "EVENT_NOT_FOUND", "EVENT_NOT_PUBLISHED", "INTERNAL_ERROR", "INVALID_CREDENTIALS", "INVALID_JSON", "METHOD_NOT_ALLOWED", "NOTIFICATION_NOT_FOUND", "ORGANIZATION_ACCESS_DENIED", "ORGANIZATION_INVITATION_EXISTS", "ORGANIZATION_INVITATION_INVALID", "ORGANIZATION_MEMBER_CHANGE_DENIED", "ORGANIZATION_MEMBER_EXISTS", "ORGANIZATION_MEMBER_NOT_FOUND", "ORGANIZATION_OWNER_TRANSFER_DENIED", "ORGANIZATION_SLUG_IN_USE", "ORGANIZER_NOT_FOUND", "PARTICIPATION_REQUIRED", "PASSWORD_RESET_TOKEN_INVALID", "POST_NOT_FOUND", "RECOVERY_EMAIL_ALREADY_BOUND", "RECOVERY_EMAIL_IN_USE", "RECOVERY_EMAIL_RATE_LIMITED", "RECOVERY_EMAIL_TOKEN_INVALID", "REGISTRATION_DUPLICATE", "REGISTRATION_NOT_FOUND", "REPLY_NOT_FOUND", "REQUEST_TOO_LARGE", "TICKET_NOT_FOUND", "TICKET_SOLD_OUT", "USER_ALREADY_EXISTS", "USER_AUTH_REQUIRED", "USER_TOKEN_INVALID", "VALIDATION_ERROR"] },
          "message": { "type": "string" }
        }
      },
...
      "CreateOrganizationRequest": { "type": "object", "additionalProperties": false, "required": ["name", "slug", "profile_name"], "properties": { "name": { "type": "string" }, "slug": { "type": "string", "pattern": "^[a-z0-9][a-z0-9-]*[a-z0-9]$" }, "profile_name": { "type": "string" }, "profile_description": { "type": "string" }, "profile_contact": { "type": "string" }, "profile_logo_url": { "type": "string", "format": "uri-reference" }, "profile_address": { "type": "string" }, "profile_website": { "type": "string", "format": "uri-reference" }, "profile_tags": { "type": "string" } } },
      "CreateOrganizationInvitationRequest": { "type": "object", "additionalProperties": false, "required": ["email", "role"], "properties": { "email": { "type": "string", "format": "email" }, "role": { "type": "string", "enum": ["admin", "editor", "checker", "finance"] } } },
      "AcceptOrganizationInvitationRequest": { "type": "object", "additionalProperties": false, "required": ["token"], "properties": { "token": { "type": "string", "minLength": 1 } } },
      "UpdateOrganizationMemberRequest": { "type": "object", "additionalProperties": false, "required": ["role"], "properties": { "role": { "type": "string", "enum": ["admin", "editor", "checker", "finance"] } } },
      "TransferOrganizationOwnerRequest": { "type": "object", "additionalProperties": false, "required": ["member_id"], "properties": { "member_id": { "type": "integer", "format": "int64", "minimum": 1 } } },
      "CreateOrganizerRequest": { "type": "object", "additionalProperties": false, "required": ["name"], "properties": { "name": { "type": "string" }, "description": { "type": "string" }, "contact": { "type": "string" }, "logo_url": { "type": "string" }, "address": { "type": "string" }, "website": { "type": "string" }, "tags": { "type": "string" } } },
      "UpdateOrganizerRequest": { "type": "object", "additionalProperties": false, "properties": { "name": { "type": "string" }, "description": { "type": "string" }, "contact": { "type": "string" }, "logo_url": { "type": "string" }, "address": { "type": "string" }, "website": { "type": "string" }, "tags": { "type": "string" } } },
...
    }
  }
}